Updateed main.py
Decode the GSM operator and signal strength received from the GSM module and print it to the console. This will later be displayed on the OLED screen.
Blink RED and BLUE on the RGB led when the program raises an OSError.
The main function will check for incoming calls with each loop iteration.